move encode pipe into separate module
