+++ libs/common/src/lib/lto-common.module.ts
@@ -6,7 +6,6 @@ import { HttpErrorComponent } from './components/http-error/http-error.component
 import { LoadingSpinnerComponent } from './components/loading-spinner/loading-spinner.component';
 import { ResponsiveTextComponent } from './components/responsive-text/responsive-text.component';
 import { LtoPipe } from './pipes/lto.pipe';
-import { EncodePipe } from './pipes/encode.pipe';
 import { TransactionLabelPipe } from './pipes/transaction-label.pipe';
 
 const commonComponents = [
@@ -15,7 +14,7 @@ const commonComponents = [
   LoadingSpinnerComponent,
   ResponsiveTextComponent
 ];
-const commonPipes = [LtoPipe, EncodePipe, TransactionLabelPipe];
+const commonPipes = [LtoPipe, TransactionLabelPipe];
 
 @NgModule({
   imports: [CommonModule, MatProgressSpinnerModule],
