+++ libs/common/src/lib/pipes/encode/encode-pipe.module.ts
@@ -1,0 +1,8 @@
+import { NgModule } from '@angular/core';
+import { EncodePipe } from './encode.pipe';
+
+@NgModule({
+  declarations: [EncodePipe],
+  exports: [EncodePipe]
+})
+export class EncodePipeModule {}
