+++ libs/common/src/lib/pipes/encode/index.ts
@@ -1,0 +1,2 @@
+export * from './encode-pipe.module';
+export * from './encode.pipe';
