+++ libs/common/src/lib/pipes/index.ts
@@ -1,3 +1,3 @@
 export * from './lto.pipe';
-export * from './encode.pipe';
+export * from './encode';
 export * from './transaction-label.pipe';
